cni: Retrieve aws-cni configuration from existing file
With IPv6 another plugin was introduced - since this plugin needs a bit
of configuration the first approach is to simply tag onto the existing
conflist. If the original aws-cni config cannot be found we revert to
the old behavior of simply using a pre-made config. On IPv6 clusters
will lead to IPv4 NAT not being available, cluster functionality itself
will work just fine though.

+++ plugins/cilium-cni/cni-install.sh
@@ -174,7 +174,12 @@ EOF
 	;;
 
 "aws-cni")
-	cat > "${CNI_CONF_NAME}" <<EOF
+  if [ -f "${CNI_CONF_DIR}/10-aws.conflist.cilium_bak" ]; then
+    jq --argjson 'ENABLE_DEBUG' "$ENABLE_DEBUG" --arg 'LOG_FILE' "$LOG_FILE" \
+          '.plugins += [{"name": "cilium", "type": "cilium-cni", "enable-debug": $ENABLE_DEBUG, "log-file": $LOG_FILE}]' \
+          "${CNI_CONF_DIR}/10-aws.conflist.cilium_bak" > "${CNI_CONF_NAME}"
+  else
+    cat > "${CNI_CONF_NAME}" <<EOF
 {
   "cniVersion": "0.3.1",
   "name": "aws-cni",
@@ -201,7 +206,8 @@ EOF
   ]
 }
 EOF
-	;;
+  fi
+  ;;
 
 *)
 	cat > "${CNI_CONF_NAME}" <<EOF
